🌱 Consolidate formatCurrency, formatPercent, formatStatNumber into lib/formatters
Move formatStatNumber, formatPercent, and formatCurrency from
stats/types.ts and valueResolvers.ts into the canonical lib/formatters.ts.
Both source files now import + re-export from the single source of truth.

The canonical formatStatNumber also handles billions and negative values
(strictly more capable, no behavior change for existing inputs).

Net reduction: ~12 lines of duplicated logic.
